+++ lib/features/password_manager/dashboard/widgets/entity_cards/api_key/api_key_grid_card.dart
@@ -61,7 +61,7 @@ class _ApiKeyGridCardState extends ConsumerState<ApiKeyGridCard> {
   Widget build(BuildContext context) {
     final item = widget.data.card.item;
     final subtitleParts = [
-      if ((_apiKey.service ?? '').isNotEmpty) _apiKey.service!,
+      if (_apiKey.service.isNotEmpty) _apiKey.service,
       if (_apiKey.environment != null) _apiKey.environment!.name,
       if (_apiKey.tokenType != null) _apiKey.tokenType!.name,
     ];