+++ lib/features/password_manager/dashboard/widgets/entity_cards/loyalty_card/loyalty_card_list_card.dart
@@ -92,7 +92,7 @@ class _LoyaltyCardListCardState extends ConsumerState<LoyaltyCardListCard> {
   Widget build(BuildContext context) {
     final item = widget.data.card.item;
     final subtitleParts = [
-      if ((_loyalty.programName ?? '').isNotEmpty) _loyalty.programName!,
+      if (_loyalty.programName.isNotEmpty) _loyalty.programName,
       if (_loyalty.barcodeType != null) _loyalty.barcodeType!.name,
     ];